[refactor] 불필요한 요소 삭제
