+++ frontend/Idle/utils/constants.jsx
@@ -5,8 +5,14 @@ export const TYPE = {
   option: "옵션",
   bill: "견적서 완성",
 };
+export const TRIM = "trim";
+export const DETAIL = "detail";
+export const COLOR = "color";
+export const OPTION = "option";
+export const BILL = "bill";
 
 export let clickedOptionPage = false;
+
 export function setClickedOptionPage() {
   clickedOptionPage = true;
 }
