+++ frontend/Idle/utils/context.jsx
@@ -1,4 +1,3 @@
 import { createContext } from "react";
 
 export const carContext = createContext();
-export const currentPageContext = createContext();
